+++ Assets/CityMap/Scripts/MapLocationManager.cs
@@ -19,7 +19,7 @@
 			MapLocation saintPaulMap = new MapLocation();
 			saintPaulMap._latLongString = "44.9504, -93.1015";
 			saintPaulMap._zoom = 13;
-			saintPaulMap._tileSize = 0.1;
+			saintPaulMap._tileSize = (float) 0.1;
 			saintPaulMap._westTiles = 2;
 			saintPaulMap._northTiles = 1;
 			saintPaulMap._eastTiles = 2;
@@ -29,7 +29,7 @@
 			MapLocation seattleMap = new MapLocation();
 			seattleMap._latLongString = "47.6038, -122.3301";
 			seattleMap._zoom = 13;
-			seattleMap._tileSize = 0.1;
+			seattleMap._tileSize = (float) 0.1;
 			seattleMap._westTiles = 2;
 			seattleMap._northTiles = 5;
 			seattleMap._eastTiles = 2;
@@ -39,7 +39,7 @@
 			MapLocation minneapolisMap = new MapLocation();
 			minneapolisMap._latLongString = "44.9773, -93.2655";
 			minneapolisMap._zoom = 13;
-			minneapolisMap._tileSize = 0.1;
+			minneapolisMap._tileSize = (float) 0.1;
 			minneapolisMap._westTiles = 1;
 			minneapolisMap._northTiles = 2;
 			minneapolisMap._eastTiles = 2;
@@ -49,7 +49,7 @@
 			MapLocation detroitMap = new MapLocation();
 			detroitMap._latLongString = "42.3487, -83.0567";
 			detroitMap._zoom = 13;
-			detroitMap._tileSize = 0.1;
+			detroitMap._tileSize = (float) 0.1;
 			detroitMap._westTiles = 2;
 			detroitMap._northTiles = 4;
 			detroitMap._eastTiles = 3;
